pyproject.toml
[tool.poetry]
name = "gradcam"
version = "0.1.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
open-clip-torch = "^2.23.0"
timm = "^0.9.9"
transformers = "^4.36.1"
pillow = "^10.1.0"
opencv-python = "^4.8.1.78"
matplotlib = "^3.8.1"
pyqt5 = "^5.15.10"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
